Customers churn
Job simulation exercise: I am working as a marketing analyst for a large telecom company. The head of marketing wants to know why customers are churning and prepare a plan to reduce customer churn. I am provided with company data and my task is to explore the data and find some insights that will help the marketing head to better strategize the marketing campaign for the next year.