Sentiment analysis- sentences
For a Toyota Social Media project, we analyzed 3 videos: a Toyota review, comparison with competitor and reliability. We extracted the comments from the youtube videos with text mining and analyzed the sentiment using 5 models: SentiStrength, Vader, BlobText, Afinn and SentiWordNet. Then we did the combined sentiment score, the average of the whole data and 2 lists, negative and positive texts , depending if their average sentiment socre was below or above 0, finally concluded an overall sentiment towards the product by Gen Z in Youtube comments.